added email config in backend

+++ backend/Repositories/Travels/IManagerRepository.cs
@@ -4,6 +4,6 @@ namespace backend.Repositories.Travels;
 
 public interface IManagerRepository
 {
-    Task<IReadOnlyCollection<TeamMemberDto>> GetTeamMembersAsync(int managerId);
-    Task<IReadOnlyCollection<TeamExpenseDto>> GetTeamExpensesAsync(int managerId, DateTime? from, DateTime? to);
+    Task<IReadOnlyCollection<TeamMemberDto>> GetTeamMembersAsync(long managerId);
+    Task<IReadOnlyCollection<TeamExpenseDto>> GetTeamExpensesAsync(long managerId, DateTime? from, DateTime? to);
 }
